refactor: move composition data types to @hyperframes/parsers
