+++ packages/parsers/src/index.ts
@@ -1,2 +1,1 @@
-// barrel — populated in U3
-export {};
+export * from "./types.js";
